requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Pillow==10.1.0
numpy==1.26.4
Flask-Cors==4.0.0